% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{AddTrailingDotIfMissing}
\alias{AddTrailingDotIfMissing}
\title{AddTrailingDotIfMissing}
\usage{
AddTrailingDotIfMissing(string = "stairway.to.heaven")
}
\arguments{
\item{string}{A string to check for a missing trailing dot.}
}
\description{
Adds a trailing dot ('.') to a string if it is missing.
}
\examples{
AddTrailingDotIfMissing(string = "stairway.to.heaven")

}
