% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{AddTrailingDotIfNonePresent}
\alias{AddTrailingDotIfNonePresent}
\title{AddTrailingDotIfNonePresent}
\usage{
AddTrailingDotIfNonePresent(string = "stairway.to.heaven")
}
\arguments{
\item{string}{The file path potentially missing the trailing slash}
}
\description{
Adds a final slash '/', if missing from a string (file path).
}
\examples{
AddTrailingDotIfNonePresent(string = "stairway.to.heaven")

}
