% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{AddTrailingSlashIfMissing}
\alias{AddTrailingSlashIfMissing}
\title{AddTrailingSlashIfMissing}
\usage{
AddTrailingSlashIfMissing(string = "stairway/to/heaven")
}
\arguments{
\item{string}{A file path to check for a missing trailing slash.}
}
\description{
Adds a trailing slash ('/') to a string if it is missing.
}
\examples{
AddTrailingSlashIfMissing(string = "stairway/to/heaven")

}
