% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{countDotOrUnderscoreSeparated}
\alias{countDotOrUnderscoreSeparated}
\title{Identify the dominant separator in a string}
\usage{
countDotOrUnderscoreSeparated(string)
}
\arguments{
\item{string}{A character string to analyze.}
}
\value{
A character string identifying the dominant separator: "dot",
"underscore", "white space", "none", or "undecided".
}
\description{
Count dots, underscores, and white spaces in a string
to guess the most prevalent separator.
}
\examples{
\dontrun{
countDotOrUnderscoreSeparated("Hello.World...")
countDotOrUnderscoreSeparated("add_translated_metadata")
countDotOrUnderscoreSeparated("add_translated.metadata")
countDotOrUnderscoreSeparated("add translated metadata")
countDotOrUnderscoreSeparated("addTranslatedMetadata")
}
}
