% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{eval_parse_kollapse}
\alias{eval_parse_kollapse}
\title{eval_parse_kollapse}
\usage{
eval_parse_kollapse(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
evaluate and parse (dyn_var_caller)
}
