% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{fix_special_characters_bash}
\alias{fix_special_characters_bash}
\title{Fix Special Characters for Bash}
\usage{
fix_special_characters_bash(path)
}
\arguments{
\item{path}{A character string representing the path to be fixed.}
}
\value{
A character string with special characters escaped for Bash.
}
\description{
This function takes a string representing a path and escapes certain special
characters to make it compatible with Bash. Specifically, it escapes spaces,
opening parentheses, and closing parentheses by placing a backslash before them.
}
\examples{
path <- "~/Dropbox (VBC)/Abel.IMBA/Data.dropbo"
fixed_path <- fix_special_characters_bash(path)
print(fixed_path) # Outputs: ~/Dropbox\ \(VBC\)/Abel.IMBA/Data.dropbo

}
