% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{FixPlotName}
\alias{FixPlotName}
\title{FixPlotName}
\usage{
FixPlotName(string = "obj@meta$alpha[[3]]", ...)
}
\arguments{
\item{string}{Input string}

\item{...}{Additional strings to concatenate after the first one}
}
\description{
FixPlotName replaces special characters in an input string (dollar-, at-, bracket-signs)
}
\examples{
FixPlotName(string = "obj at meta$alpha[[3]]")
}
