% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{FixUnderscores}
\alias{FixUnderscores}
\title{FixUnderscores}
\usage{
FixUnderscores(string = "stairway__to_heaven_", trimFinal = TRUE)
}
\arguments{
\item{string}{The string potentially containing repeated underscores.}

\item{trimFinal}{Remove the final underscore? Default: TRUE.}
}
\value{
A string with repeated underscores condensed and trailing underscores optionally removed.
}
\description{
FixUnderscores removes multiple consecutive underscores from a string and optionally trims a trailing underscore.
}
\examples{
FixUnderscores(string = "stairway__to_heaven_")
}
