% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringParser.R
\name{flag.name_value}
\alias{flag.name_value}
\title{flag.name_value}
\usage{
flag.name_value(toggle, Separator = "_")
}
\arguments{
\item{toggle}{Binary variable}

\item{Separator}{Separator, Default: '_'}
}
\description{
Returns the name and its value, if its not FALSE.
}
