% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{flag.nameiftrue}
\alias{flag.nameiftrue}
\title{flag.nameiftrue}
\usage{
flag.nameiftrue(toggle, prefix = NULL, suffix = NULL, name.if.not = "")
}
\arguments{
\item{toggle}{Binary variable}

\item{prefix}{prefix added before the string, Default: NULL}

\item{suffix}{suffix added after the string, Default: NULL}

\item{name.if.not}{Alternative name., Default: ''}
}
\description{
Returns the name and its value, if its TRUE.
}
