% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{flag.names_list.all.new}
\alias{flag.names_list.all.new}
\title{flag.names_list.all.new}
\usage{
flag.names_list.all.new(pl = p.hm)
}
\arguments{
\item{pl}{List of parameters, Default: p.hm}
}
\description{
Returns the name and value of each element in a list of parameters.
}
