% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{FormatAsExcelLink}
\alias{FormatAsExcelLink}
\title{FormatAsExcelLink}
\usage{
FormatAsExcelLink(site_name, site_url)
}
\arguments{
\item{site_name}{Text shown.}

\item{site_url}{Hyperlink url.}
}
\description{
FormatAsExcelLink
}
\examples{
FormatAsExcelLink(site_name = c("Zero Hedge", "Free Software Foundation"), site_url = c("https://www.zerohedge.com", "https://www.fsf.org"))
}
