% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{HasNames}
\alias{HasNames}
\title{Check if Vector Has Names}
\usage{
HasNames(x)
}
\arguments{
\item{x}{Vector to check.}
}
\value{
Logical indicating if \code{x} has names.
}
\description{
Check if Vector Has Names
}
\examples{
HasNames(c(a = 1, b = 2))
HasNames(1:3)
}
