% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{iprint}
\alias{iprint}
\title{iprint}
\usage{
iprint(...)
}
\arguments{
\item{...}{Values to collapse consecutively with spaces.}
}
\description{
A more intelligent printing function that collapses any variable passed to it by white spaces.
}
\examples{
iprint("Hello ", "you ", 3, ", ", 11, " year old kids.")
}
