% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{irequire}
\alias{irequire}
\title{irequire}
\usage{
irequire(package)
}
\arguments{
\item{package}{Packagename to load}
}
\description{
Load a package. If it does not exist, try to install it from CRAN.
}
\examples{
irequire(gtools)
}
