% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{is.character.or.NULL}
\alias{is.character.or.NULL}
\title{Check if Input is Character or NULL}
\usage{
is.character.or.NULL(x)
}
\arguments{
\item{x}{The input to check.}
}
\value{
Returns \code{TRUE} if \code{x} is either a character vector or \code{NULL}, otherwise \code{FALSE}.
}
\description{
\code{is.character.or.NULL()} verifies if the provided input is either a character vector or NULL.
}
\examples{
is.character.or.NULL(NULL)      # TRUE
is.character.or.NULL("example") # TRUE
is.character.or.NULL(123)       # FALSE

}
