% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{knl}
\alias{knl}
\title{Collapse and paste by newline (\verb{\\n}) preceded by a white space}
\usage{
knl(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by white space
}
\examples{
knl("A", 1:2, "end")
}
