% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kpnl}
\alias{kpnl}
\title{Collapse and paste by newline (\verb{\\n}) preceded by a white space}
\usage{
kpnl(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by newline (\verb{\\n}) preceded by a white space
}
\examples{
kpnl("A", 1:2, "end")
}
