% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kpp}
\alias{kpp}
\title{Collapse by point}
\usage{
kpp(..., trim = FALSE)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{trim}{Trim empty parameters resulting in double dots? Up to 3; FALSE by default}
}
\description{
Collapse by point
}
