% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kppc}
\alias{kppc}
\title{Collapse and paste by comma (and white space)}
\usage{
kppc(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by comma and white space (\verb{, })
}
\examples{
kppc("A", 1:2, "end")
}
