% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kppd}
\alias{kppd}
\title{Collapse and paste by dash}
\usage{
kppd(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by dash (\code{-})
}
\examples{
kppd("A", 1:2, "end")
}
