% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kpps}
\alias{kpps}
\title{Collapse by (forward) slash}
\usage{
kpps(..., trim = FALSE)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{trim}{Trim empty parameters resulting in double slashes? Up to 3; FALSE by default}
}
\description{
Collapse by (forward) slash
}
