% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kwNames}
\alias{kwNames}
\title{Paste Elements With Names}
\usage{
kwNames(x = c(a = 1, b = 2), sep1 = ":", sep2 = " | ")
}
\arguments{
\item{x}{A named vector. Default is \code{c('a' = 1, 'b' = 2)}.}

\item{sep1}{A character string to separate the names from the elements. Default is ":".}

\item{sep2}{A character string to separate the name-element pairs in the
resulting string. Default is " ".}
}
\description{
This function takes a named vector and returns a string where each element is pasted
with its name. Elements are separated by a specified string, and name-element pairs are also
separated by a specified string. The default named vector is \code{c('a' = 1, 'b' = 2)}.
}
\examples{
kwNames(c("a" = 1, "b" = 2))
}
