% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{message2}
\alias{message2}
\title{Message without collapsing}
\usage{
message2(vec)
}
\arguments{
\item{vec}{A character vector to be printed.}
}
\description{
This function prints a message for each element in a character vector, instead of
collapsing them into a single line as done by, \code{message()}.
}
\examples{
message2(c("Hello", "world", "!", "I", "am", "here"))
}
