% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{messageWS}
\alias{messageWS}
\title{messagews}
\usage{
messageWS(..., collapse = " ")
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}
}
\description{
message pasted with white space
}
\examples{
iprint("Hello ", "you ", 3, ", ", 11, " year old kids.")
}
