% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{param.list.flag}
\alias{param.list.flag}
\title{param.list.flag}
\usage{
param.list.flag(par)
}
\arguments{
\item{par}{Parameter value.}
}
\description{
Returns the name and value of each element in a list of parameters.
}
\examples{
umap_min_dist <- 0.1
param.list.flag(umap_min_dist)

}
