% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{params.2.fname}
\alias{params.2.fname}
\title{Convert Named Parameters to Filename}
\usage{
params.2.fname(..., sep = ".", collapse = "_")
}
\arguments{
\item{...}{Named parameters to be converted. Default: None.}

\item{sep}{A string to separate parameter names and their values. Default: ".".}

\item{collapse}{A string to separate different parameters in the output string. Default: "_".}
}
\value{
A character string that represents the combined parameter names and values, separated
by the specified \code{sep} and \code{collapse} characters.
}
\description{
This function takes named parameters and converts them into a filename string with
specified separators and collapse characters. It excludes any parameters with NULL values.
}
\examples{
params.2.fname(aa = 1, cc = 2, d = NULL, sep = ".", collapse = "_")
# Returns "aa.1_cc.2"
}
