% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ParseDirPath}
\alias{ParseDirPath}
\title{ParseDirPath}
\usage{
ParseDirPath(...)
}
\arguments{
\item{...}{The set of strings (character vectors) to be parsed into a file path, and potentially
having Double Slashes, potentially missing a trailing slash.}
}
\description{
ParseDirPath pastes elements by slash, then removes Double Slashes '//' from a
string and adds a final '/' if missing from a file path.
}
\examples{
ParseDirPath(string = "stairway///to/heaven")
}
