% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ParseFullFilePath}
\alias{ParseFullFilePath}
\title{Parse Full File Path}
\usage{
ParseFullFilePath(path, file_name, extension)
}
\arguments{
\item{path}{The directory path. If not provided, only file name and extension are used.
Default: NULL.}

\item{file_name}{The name of the file. Clean-up operations are applied to remove special
characters and repeated dots. Default: Empty string.}

\item{extension}{The file extension. If provided, it is appended to the file name with a
preceding dot. Clean-up operations remove any initial dots. Default: NULL.}
}
\value{
A string representing the full file path.
}
\description{
Constructs a full file path by combining a path, file name, and extension. It applies
string clean-up operations to each component and ensures proper formatting.
}
\examples{
ParseFullFilePath(path = "home/user/docs/", file_name = "report@final", extension = ".txt")
ParseFullFilePath(file_name = "report", extension = "txt")
}
