% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{parseParamStringWNames}
\alias{parseParamStringWNames}
\title{Parse Parameter String with Names}
\usage{
parseParamStringWNames(named.vec, sep1 = ": ", sep2 = " | ")
}
\arguments{
\item{named.vec}{A named vector to be parsed. Default: \code{NA}.}

\item{sep1}{A string separator for odd elements. Default: \code{": "}.}

\item{sep2}{A string separator for even elements. Default: \code{" | "}.}
}
\value{
A single string with intermingled names and values from the named vector.
}
\description{
This function parses a named vector and intermingles the names and values
into a single string, with specified separators for the odd and even elements.
}
\examples{
named.vec <- c(ULm = "15", DLm = "67", Matm = "33", `EN-Lineage.m` = "21")
parseParamStringWNames(named.vec)
# "ULm: 15 | DLm: 67 | Matm: 33 | EN-Lineage.m: 21"

}
