% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{paste_w_names}
\alias{paste_w_names}
\title{Paste Elements With Names}
\usage{
paste_w_names(
  named_vec = c(a = 1, b = 2),
  separator_names = ": ",
  separator_elements = " | "
)
}
\arguments{
\item{named_vec}{A named vector. Default is \code{c('a' = 1, 'b' = 2)}.}

\item{separator_names}{A character string to separate the names from the elements. Default is ":".}

\item{separator_elements}{A character string to separate the name-element pairs in the resulting string. Default is " ".}
}
\description{
This function takes a named vector and returns a string where each element is pasted with its name. Elements are separated by a specified string, and name-element pairs are also separated by a specified string. The default named vector is \code{c('a' = 1, 'b' = 2)}.
}
