% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringParser.R
\name{percentile2value}
\alias{percentile2value}
\title{percentile2value}
\usage{
percentile2value(
  distribution,
  percentile = 0.95,
  FirstValOverPercentile = TRUE
)
}
\arguments{
\item{distribution}{A numeric vector}

\item{percentile}{percentile, Default: 0.95}

\item{FirstValOverPercentile}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Calculate what is the actual value of the N-th percentile in a distribution or set of numbers. Useful for calculating cutoffs, and displaying them by whist()s "vline" paramter.
}
