% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ppc}
\alias{ppc}
\title{Paste by a comma followed by a white space}
\usage{
ppc(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Paste by a comma followed by a white space
}
