% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ppipe}
\alias{ppipe}
\title{Paste by pipe (|) and white space around it}
\usage{
ppipe(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Paste by pipe (|) and white space around it
}
