% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveDoubleDot}
\alias{RemoveDoubleDot}
\title{RemoveDoubleDot}
\usage{
RemoveDoubleDot(string = "stairway...to.heaven.")
}
\arguments{
\item{string}{The file path potentially having Double Dot}
}
\description{
RemoveDoubleDot removes multiple consecutive slashes (e.g. '..') from a string (file path). Also works for 2,3 consecutive slashes
}
\examples{
RemoveDoubleDot(string = "stairway..to...heaven....") # replace by a single .
}
