% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveFinalDot}
\alias{RemoveFinalDot}
\title{RemoveFinalDot}
\usage{
RemoveFinalDot(string = "stairway.to.heaven.")
}
\arguments{
\item{string}{The file path potentially having Final Dot}
}
\description{
RemoveFinalDot removes the final slash from a string
}
\examples{
RemoveFinalDot(string = "stairway..to...heaven...")
}
