% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveFinalSlash}
\alias{RemoveFinalSlash}
\title{RemoveFinalSlash}
\usage{
RemoveFinalSlash(string = "stairway/to/heaven//")
}
\arguments{
\item{string}{The file path potentially having Final Slash}
}
\description{
RemoveFinalSlash removes the final slash from a string
}
\examples{
RemoveFinalSlash(string = "stairway//to///heaven//")
}
