% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveFinalUnderscores}
\alias{RemoveFinalUnderscores}
\title{RemoveFinalUnderscores}
\usage{
RemoveFinalUnderscores(string)
}
\arguments{
\item{string}{The string (file path) potentially having a final slash.}
}
\value{
A string with the final slash removed.
}
\description{
RemoveFinalUnderscores removes the final slash(es) from a string (file path).
}
\examples{
RemoveFinalUnderscores(string = "path/to/folder/")
}
