% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveInitialDot}
\alias{RemoveInitialDot}
\title{RemoveInitialDot}
\usage{
RemoveInitialDot(string)
}
\arguments{
\item{string}{The string potentially having an initial dot.}
}
\value{
A string with the initial dot removed.
}
\description{
RemoveInitialDot removes the initial dot from a string.
}
\examples{
RemoveInitialDot(string = ".example...")
}
