% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveTrailingDots}
\alias{RemoveTrailingDots}
\title{RemoveTrailingDots}
\usage{
RemoveTrailingDots(string = "...stairway.to..heaven.")
}
\arguments{
\item{string}{The string potentially having 1-3 dots at the beginnig or end}
}
\description{
RemoveTrailingDots removes the trailing dots from a string
}
\examples{
RemoveTrailingDots(string = "...stairway.to..heaven.")
}
