% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveWhitespaces}
\alias{RemoveWhitespaces}
\title{RemoveWhitespaces}
\usage{
RemoveWhitespaces(string, replacement = "")
}
\arguments{
\item{string}{The string potentially containing whitespace.}

\item{replacement}{The string to replace whitespace with. Default: ''.}
}
\value{
A string with all whitespace replaced by \code{replacement}.
}
\description{
RemoveWhitespaces removes all whitespace characters from a string or replaces them with a specified value.
}
\examples{
RemoveWhitespaces(string = "path   To    Folder")
}
