% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceFinalDot}
\alias{ReplaceFinalDot}
\title{ReplaceFinalDot}
\usage{
ReplaceFinalDot(string)
}
\arguments{
\item{string}{The file path potentially having Final Dot}
}
\description{
ReplaceFinalDot removes the final dot from a string
}
\examples{
ReplaceFinalDot(string = "stairway..to...heaven...")

}
