% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceFinalSlash}
\alias{ReplaceFinalSlash}
\title{ReplaceFinalSlash}
\usage{
ReplaceFinalSlash(string)
}
\arguments{
\item{string}{The string (file path) potentially having a final slash.}
}
\value{
A string with the final slash removed.
}
\description{
ReplaceFinalSlash removes the final slash from a string (file path).
}
\examples{
ReplaceFinalSlash(string = "path/to/folder/")
}
