% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceInitialDot}
\alias{ReplaceInitialDot}
\title{ReplaceInitialDot}
\usage{
ReplaceInitialDot(string)
}
\arguments{
\item{string}{The string potentially having an initial dot.}
}
\value{
A string with the initial dot removed.
}
\description{
ReplaceInitialDot removes the initial dot from a string.
}
\examples{
ReplaceInitialDot(string = ".example...")
}
