% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceRepeatedDots}
\alias{ReplaceRepeatedDots}
\title{ReplaceRepeatedDots}
\usage{
ReplaceRepeatedDots(string)
}
\arguments{
\item{string}{The string (file name or path) potentially having multiple dots}
}
\description{
ReplaceRepeatedDots removes multiple consecutive slashes (e.g. '..') from a string (file path).
}
\examples{
ReplaceRepeatedDots(string = "stairway..to...heaven....") # replace by a single dot.

}
