% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceRepeatedSlashes}
\alias{ReplaceRepeatedSlashes}
\title{ReplaceRepeatedSlashes}
\usage{
ReplaceRepeatedSlashes(string)
}
\arguments{
\item{string}{The string (file path) potentially having repeated slashes.}
}
\value{
A string with repeated slashes replaced by a single slash.
}
\description{
ReplaceRepeatedSlashes replaces multiple consecutive slashes with a single slash.
}
\examples{
ReplaceRepeatedSlashes(string = "path//to//folder")
}
