% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceRepeatedWhitespaces}
\alias{ReplaceRepeatedWhitespaces}
\title{ReplaceRepeatedWhitespaces}
\usage{
ReplaceRepeatedWhitespaces(string, replacement = " ")
}
\arguments{
\item{string}{The string potentially containing repeated whitespace.}

\item{replacement}{The string to replace repeated whitespace with. Default: ' '.}
}
\value{
A string with repeated whitespace replaced by \code{replacement}.
}
\description{
ReplaceRepeatedWhitespaces collapses multiple consecutive whitespace characters into a single replacement.
}
\examples{
ReplaceRepeatedWhitespaces(string = "path   to    folder")
}
