% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceSpecialCharacters}
\alias{ReplaceSpecialCharacters}
\title{ReplaceSpecialCharacters}
\usage{
ReplaceSpecialCharacters(string = "obj@meta$alpha[[3]]")
}
\arguments{
\item{string}{The string, e.g variable subset: "obj@meta$alpha[\link{3}]"}
}
\description{
ReplaceSpecialCharacters replaces '[]$@()' with dots
}
\examples{
# ReplaceSpecialCharacters('string with at bracked and dollar signs')
}
