% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceWeirdCharacters}
\alias{ReplaceWeirdCharacters}
\title{ReplaceWeirdCharacters}
\usage{
ReplaceWeirdCharacters(string = "obj@meta$alpha[[3]]")
}
\arguments{
\item{string}{The string, e.g variable subset: "obj at meta$alpha[\link{3}]"}
}
\description{
ReplaceWeirdCharacters replaces '[]$ at and ()' with dots
}
\examples{
# ReplaceWeirdCharacters('string with at bracked and dollar signs')
}
