% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{sppp}
\alias{sppp}
\title{Simplified Paste by point}
\usage{
sppp(..., make.names = FALSE)
}
\arguments{
\item{...}{Multiple simple string variables to parse.}

\item{make.names}{Should make.names applied to the concatenated string? Default is FALSE.}
}
\description{
Simplified Paste by point
}
\examples{
sppp("A", 1:2, "end", "", NULL); kpp("A", 1:2, "end", "", NULL)
}
