% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{spps}
\alias{spps}
\title{Simplified Paste by fwd slash}
\usage{
spps(..., make.names = FALSE)
}
\arguments{
\item{...}{Multiple simple variables to parse.}

\item{make.names}{Should make.names applied to the concatenated string? Default is FALSE.}
}
\description{
Simplified Paste by fwd slash
}
\examples{
spps("A", 1:2, "end", "", NULL); kpps("A", 1:2, "end", "", NULL)
}
