% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{stopif}
\alias{stopif}
\title{Stop execution if condition is TRUE}
\usage{
stopif(...)
}
\arguments{
\item{...}{Logical conditions to be checked. Each condition must evaluate to a logical vector.
Default: none. Named arguments will be used as error messages.}
}
\value{
The function stops execution if any condition evaluates to \code{TRUE}.
}
\description{
The \code{stopif()} function stops the execution if the condition is \code{TRUE}. It is the opposite of
\code{stopifnot()}, which stops if the condition is not \code{TRUE}. This function is useful to increase
clarity in the code by removing double negations.
}
\examples{
try(stopif(6 < 4, 6 > 5), silent = TRUE)
try(stopif("custom message" = 6 > 5), silent = TRUE)

}
