% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{substitute_deparse}
\alias{substitute_deparse}
\title{Get Object Name as String}
\usage{
substitute_deparse(x)
}
\arguments{
\item{x}{The object whose name you want to capture.}
}
\value{
Returns the name of \code{obj} as a string.
}
\description{
\code{substitute_deparse()} captures its input expression and returns the deparsed expression as a
character string.
Replace \verb{deparse\\s*\\(\\s*substitute\\s*\\(([^()]+)\\)\\s*\\)} to \verb{substitute_deparse($1)}, then
\verb{substitute\\s*\\(([^()]+)\\)\\s*\\)} to the same.
}
\examples{
my_var <- 10
substitute_deparse(my_var) # "my_var"

}
