% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{testNumericCompatible}
\alias{testNumericCompatible}
\title{Test if a Variable is Inherently Numeric ('0.1' as numeric)}
\usage{
testNumericCompatible(x)
}
\arguments{
\item{x}{The variable to be tested.}
}
\value{
A logical value indicating whether the input is inherently numeric.
}
\description{
This function checks if a given variable is inherently numeric. It returns TRUE if the variable
can be converted to a numeric value without loss of information and is not inherently a
character string, otherwise it returns FALSE.
}
\examples{
testNumericCompatible(0.1) # Should return TRUE
testNumericCompatible("0.1") # Should return TRUE
testNumericCompatible("apple") # Should return FALSE
testNumericCompatible("arma.0.1") # Should return FALSE

}
