% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{toDotSeparated}
\alias{toDotSeparated}
\title{Convert String to Dot Separated Name}
\usage{
toDotSeparated(input_string, toclipboard = TRUE)
}
\arguments{
\item{input_string}{A character string in camelCase or underscore_separated format to be converted.
Default: No default value, a string must be provided.}

\item{toclipboard}{Copy to clipboard? Default: TRUE}
}
\value{
Invisibly, a lowercase character string in dot-separated format.
}
\description{
Converts a string from camelCase or underscore_separated format to dot.separated.name format.
Inserts dots before each uppercase letter (except if it's the first character) or replaces underscores with dots,
and then converts the entire string to lowercase.
}
\examples{
toDotSeparated("plotMetadataCorHeatMap")
toDotSeparated("plot_Metadata_Cor_HeatMap")
}
