% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{warnif}
\alias{warnif}
\title{Issue a warning if condition is TRUE}
\usage{
warnif(...)
}
\arguments{
\item{...}{Logical conditions to be checked. Each condition must evaluate to a logical vector.
Named arguments will use the name as the warning message.}
}
\value{
The function returns \code{invisible()}, but issues warnings for each condition that evaluates
to \code{TRUE}.
}
\description{
The \code{warnif()} function issues a warning if the condition is \code{TRUE}. It is the opposite of
\code{warnifnot()}, which warns if the condition is not \code{TRUE}. This function is useful for issuing
warnings when a certain condition is met.
}
\examples{
warnif(6 > 4, 6 > 5)
warnif("custom message" = 6 > 5)

}
