% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{warnifnot}
\alias{warnifnot}
\title{Issue warnings if conditions are not TRUE}
\usage{
warnifnot(...)
}
\arguments{
\item{...}{Logical conditions to be checked. Each condition must evaluate to a logical vector.
Named arguments will use the name as the warning message.}
}
\value{
The function returns \code{invisible()}, but issues warnings for each condition that evaluates
to \code{FALSE} or contains \code{NA}.
}
\description{
The \code{warnifnot()} function checks whether each condition passed to it is \code{TRUE}. If any
condition is not met, a warning is issued but execution continues. This is similar to
\code{stopifnot()}, which throws an error and halts execution, but \code{warnifnot()} only issues a
warning, allowing the program to proceed.
}
\examples{
warnifnot(6 < 4, 6 > 5)
warnifnot("custom message" = 6 < 4)

}
