% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ww.break.lines}
\alias{ww.break.lines}
\title{break.lines for plot titles}
\usage{
ww.break.lines(char.vec = kppd(LETTERS), max.char = 50)
}
\arguments{
\item{char.vec}{A long sentence}

\item{max.char}{Max characters per line}
}
\description{
break.lines for plot titles
}
\examples{
ww.break.lines()
}
