% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{flag.names_list}
\alias{flag.names_list}
\title{flag.names_list}
\usage{
flag.names_list(ls, sep_name_val = "_", sep_elem = "-")
}
\arguments{
\item{ls}{List of parameters (name, value).}

\item{sep_name_val}{Separator name-2-value, Default: "_"}

\item{sep_elem}{Separator between elements, Default: "-"}
}
\description{
Returns the name and value of each element in a list of parameters.
}
\examples{
parameters <- list(method = "pearson", dimensions = 20)
flag.names_list(parameters)

}
