% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{flag.names_list}
\alias{flag.names_list}
\title{flag.names_list}
\usage{
flag.names_list(par)
}
\arguments{
\item{par}{A list element e.g.: p$umap}
}
\description{
Returns the name and value of each element in a list of parameters.
}
\examples{
# flag.names_list(par = p$'umap.n_neighbors')

}
