% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Negation of the \verb{in} (w. grapes) Operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{A vector of values to be matched.}

\item{y}{A vector of values to be matched against.}
}
\value{
A logical vector indicating if elements in \code{x} are not present in \code{y}.
}
\description{
\verb{\%!in\%} is used to test if elements of one vector are not present in another vector.
It is the negation of the \code{\%in\%} operator. This operator returns \code{TRUE} for elements
of \code{x} that are not in \code{y}.
}
\examples{
c(1, 2, 3) \%!in\% c(2, 4, 6)
# [1]  TRUE FALSE  TRUE
}
