% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{idate}
\alias{idate}
\title{Parse current date, dot separated.}
\usage{
idate(Format = c("\%Y.\%m.\%d_\%H.\%M", "\%Y.\%m.\%d_\%Hh")[2])
}
\arguments{
\item{Format}{Date format. Default: c("\%Y.\%m.\%d_\%H.\%M", "\%Y.\%m.\%d_\%Hh")\link{2}}
}
\value{
A character string of the current date/time formatted according
to \code{Format}.
}
\description{
Returns the current system date and time formatted as a character
string. The default format uses dot separated components, but any
format recognised by \link[base:format]{base::format} can be supplied.
}
\examples{
idate()
idate("\%Y-\%m-\%d")
}
