% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringParser.R
\name{idate}
\alias{idate}
\title{Parse current date, dot separated.}
\usage{
idate(Format = c("\%Y.\%m.\%d_\%H.\%M", "\%Y.\%m.\%d_\%Hh")[2])
}
\arguments{
\item{Format}{PARAM_DESCRIPTION, Default: c("\%Y.\%m.\%d_\%H.\%M", "\%Y.\%m.\%d_\%Hh")\link{2}}
}
\description{
Parse current date, dot separated.
}
