% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ifExistsAndTrue}
\alias{ifExistsAndTrue}
\title{ifExistsAndTrue}
\usage{
ifExistsAndTrue(varname = "pi")
}
\arguments{
\item{varname}{Name of the variable}
}
\description{
Checks if a variable is defined, and its value is TRUE, else returns FALSE, and
prints a message.
}
\examples{
ifExistsAndTrue("pi"); ifExistsAndTrue("pi22")

}
