% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ifExistsElse}
\alias{ifExistsElse}
\title{ifExistsElse}
\usage{
ifExistsElse(varname, alternative = "define an alternative", v = F)
}
\arguments{
\item{varname}{Name of the variable}

\item{alternative}{Alternative value to return if the variable is not defined}

\item{v}{Print messages. Default is FALSE.}
}
\description{
Checks if a variable is defined, else returns an alternative value.
}
\examples{
ifExistsAndTrue("pi"); ifExistsAndTrue("pi22")

}
