% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{imessage}
\alias{imessage}
\title{imessage}
\usage{
imessage(..., collapse = " ")
}
\arguments{
\item{...}{Elements to be concatenated. Each element is converted to a string and
collapsed in order.}

\item{collapse}{String inserted between elements when collapsing. Default: " "}
}
\description{
A variant of message() pasting with white space, sibling of iprint().
}
\examples{
imessage("Hello", "you", 3, ",", 11, "year old kids.")
}
