% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kpipe}
\alias{kpipe}
\title{Collapse and paste by pipe (|) and white space}
\usage{
kpipe(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by white space
}
\examples{
kpipe("A", 1:2, "end")
}
