% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kpps}
\alias{kpps}
\title{Collapse and paste by (forward) slash}
\usage{
kpps(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by forward slash (\code{/})
}
\examples{
kpps("A", 1:2, "end")
}
