% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kppu}
\alias{kppu}
\title{Collapse by underscore}
\usage{
kppu(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by underscore
}
\examples{
kppu("A", 1:2, "end")
}
