% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{kppws}
\alias{kppws}
\title{Collapse and paste by white space}
\usage{
kppws(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Collapse by white space (\verb{ })
}
\examples{
kppws("A", 1:2, "end")
}
