% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{param.list.2.fname}
\alias{param.list.2.fname}
\title{param.list.2.fname}
\usage{
param.list.2.fname(ls.of.params = p)
}
\arguments{
\item{ls.of.params}{List of parameters, Default: p}
}
\description{
Take a list of parameters and parse a string from their names and values.
}
