% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringParser.R
\name{param.list.flag}
\alias{param.list.flag}
\title{param.list.flag}
\usage{
param.list.flag(par = p$umap.min_dist)
}
\arguments{
\item{par}{parameter, Default: p$umap.min_dist}
}
\description{
Returns the name and value of each element in a list of parameters.
}
