% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{parsepvalue}
\alias{parsepvalue}
\title{parsepvalue}
\usage{
parsepvalue(pvalue = 0.01, digits = 2, brackets = FALSE, prefix = FALSE)
}
\arguments{
\item{pvalue}{pvalue to parse. Default: 0.01}

\item{digits}{Number of digits to keep. Default: 2}

\item{brackets}{Whether to enclose the result in brackets. Default: FALSE}
}
\description{
Parse p-value from a number to a string.
}
