% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringParser.R
\name{parsepvalue}
\alias{parsepvalue}
\title{parsepvalue}
\usage{
parsepvalue(pvalue = 0.01)
}
\arguments{
\item{pvalue}{pvalue to parse, Default: 0.01}
}
\description{
Parse p-value from a number to a string.
}
