% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{PasteDirNameFromFlags}
\alias{PasteDirNameFromFlags}
\title{PasteDirNameFromFlags}
\usage{
PasteDirNameFromFlags(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Paste a dot (point) separated string from a list of inputs (that can be empty), and clean up the output string from dot multiplets (e.g: ..).
}
