% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{PasteOutdirFromFlags}
\alias{PasteOutdirFromFlags}
\title{PasteOutdirFromFlags}
\usage{
PasteOutdirFromFlags(path = "~/Dropbox/Abel.IMBA/AnalysisD", ...)
}
\arguments{
\item{path}{path, Default: '~/Dropbox/Abel.IMBA/AnalysisD'}

\item{...}{Multiple simple variables to parse.}
}
\description{
Paste OutDir from (1) a path and (2) a from a list of inputs (that can be empty), and clean up the output string from dot and forward slash multiplets (e.g: ..).
}
