% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveFinalUnderscores}
\alias{RemoveFinalUnderscores}
\title{RemoveFinalUnderscores}
\usage{
RemoveFinalUnderscores(string)
}
\arguments{
\item{string}{The string potentially ending with an underscore.}
}
\value{
A string with the final underscore removed.
}
\description{
RemoveFinalUnderscores removes trailing underscore(s) from a string.
}
\examples{
RemoveFinalUnderscores(string = "path_to_folder_")
}
