% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveTrailingDots}
\alias{RemoveTrailingDots}
\title{RemoveTrailingDots}
\usage{
RemoveTrailingDots(string = "...stairway.to..heaven.")
}
\arguments{
\item{string}{The string potentially having unnecessary dots at the beginning or end.}
}
\value{
A string with the trailing dots removed.
}
\description{
RemoveTrailingDots removes dots at the beginning and end of a string.
}
\examples{
RemoveTrailingDots(string = "...stairway.to..heaven.")
}
