% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{RemoveWhitespaces}
\alias{RemoveWhitespaces}
\title{RemoveWhitespaces}
\usage{
RemoveWhitespaces(string, replacement = "")
}
\arguments{
\item{string}{The string (file path) potentially having repeated slashes.}

\item{replacement}{The string to replace the white spaces with. Default: ''.}
}
\value{
A string with repeated slashes replaced by a single slash.
}
\description{
RemoveWhitespaces replaces any nr of white spaces.
}
\examples{
RemoveWhitespaces(string = "path   To    Folder")
}
