% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceRepeatedDots}
\alias{ReplaceRepeatedDots}
\title{ReplaceRepeatedDots}
\usage{
ReplaceRepeatedDots(string)
}
\arguments{
\item{string}{A string (file name or path) that may contain repeated dots/periods}
}
\description{
ReplaceRepeatedDots collapses multiple consecutive dots (periods) in a string into a single dot.
}
\examples{
ReplaceRepeatedDots(string = "stairway..to...heaven....") # "stairway.to.heaven."

}
