% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceRepeatedUnderscores}
\alias{ReplaceRepeatedUnderscores}
\title{ReplaceRepeatedUnderscores}
\usage{
ReplaceRepeatedUnderscores(string)
}
\arguments{
\item{string}{The string (file path) potentially having repeated underscores.}
}
\value{
A string with repeated underscores replaced by a single underscore.
}
\description{
ReplaceRepeatedUnderscores replaces multiple consecutive underscores with a single underscore.
}
\examples{
ReplaceRepeatedUnderscores(string = "path__to__folder")
}
