% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceRepeatedWhitespaces}
\alias{ReplaceRepeatedWhitespaces}
\title{ReplaceRepeatedWhitespaces}
\usage{
ReplaceRepeatedWhitespaces(string)
}
\arguments{
\item{string}{The string (file path) potentially having repeated slashes.}
}
\value{
A string with repeated slashes replaced by a single slash.
}
\description{
ReplaceRepeatedWhitespaces replaces multiple consecutive white spaces with a single one.
}
\examples{
ReplaceRepeatedUnderscores(string = "path   to    folder")
}
