% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceSpecialCharacters}
\alias{ReplaceSpecialCharacters}
\title{ReplaceSpecialCharacters}
\usage{
ReplaceSpecialCharacters(string = "obj@meta$alpha[[3]]")
}
\arguments{
\item{string}{The string potentially having special characters.}
}
\value{
A string with special characters replaced by dots.
}
\description{
ReplaceSpecialCharacters replaces special characters '[]$@()' with dots.
}
\examples{
ReplaceSpecialCharacters(string = "obj@meta$alpha[[3]]")
}
