% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{ReplaceSpecialCharacters}
\alias{ReplaceSpecialCharacters}
\title{ReplaceSpecialCharacters}
\usage{
ReplaceSpecialCharacters(
  string = "obj@meta$alpha[[3]]",
  replacement = ".",
  remove_dots = F
)
}
\arguments{
\item{string}{The string potentially having special characters.}

\item{replacement}{The character to replace special characters with.}

\item{remove_dots}{If TRUE, all dots are removed from the string (overwrites if replacement is a dot).}
}
\value{
A string with special characters replaced by dots.
}
\description{
ReplaceSpecialCharacters replaces special characters '[]$@()' with dots.
}
\examples{
ReplaceSpecialCharacters(string = "obj@meta$alpha[[3]]")
}
