% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{sppp}
\alias{sppp}
\title{Simplified Paste by point}
\usage{
sppp(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Simplified Paste by point
}
\examples{
sppp("A", 1:2, "end", "", NULL); kpp("A", 1:2, "end", "", NULL)
}
