% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{spps}
\alias{spps}
\title{Simplified Paste by fwd slash}
\usage{
spps(...)
}
\arguments{
\item{...}{Multiple simple variables to parse.}
}
\description{
Simplified Paste by fwd slash
}
\examples{
spps('Apples//are///sweet//')
}
