% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{substitute_deparse}
\alias{substitute_deparse}
\title{Get Object Name as String}
\usage{
substitute_deparse(x)
}
\arguments{
\item{x}{The object whose name you want to capture.}
}
\value{
Returns the name of \code{obj} as a string.
}
\description{
\code{get_object_name()} captures the name of an input object and returns it as a string.
}
\examples{
my_var <- 10
get_object_name(my_var)  # "my_var"

}
