% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{toCamelCase}
\alias{toCamelCase}
\title{Convert a String to camelCase}
\usage{
toCamelCase(
  input_string,
  estimated_separator = countDotOrUnderscoreSeparated(input_string),
  toclipboard = TRUE
)
}
\arguments{
\item{input_string}{A character string to be converted to camelCase. The function expects a
string where words are separated by dots. There is no default value for this parameter;
a string must be provided.}

\item{estimated_separator}{A character string representing the separator used in the input string.
Default: countDotOrUnderscoreSeparated(input_string)}

\item{toclipboard}{Copy to clipboard? Default: TRUE}
}
\value{
A character string converted to camelCase.
}
\description{
This function takes a string as input and converts it to camelCase format.
It splits the string into words using dots as separators, capitalizes the first letter of
each word (except the first word), and then concatenates them back together.
}
\examples{
toCamelCase("plot.metadata.cor.heatMap")
toCamelCase("plot_metadata_cor_heat_map")
}
