% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{toDotSeparated}
\alias{toDotSeparated}
\title{Convert String to Dot Separated Name}
\usage{
toDotSeparated(input_string, toclipboard = FALSE)

toDotSeparated(input_string, toclipboard = FALSE)
}
\arguments{
\item{input_string}{A character string in camelCase or underscore_separated format to be converted.
Default: No default value, a string must be provided.}

\item{toclipboard}{Copy to clipboard? Default: FALSE.}
}
\value{
A character string converted to dot-separated format. The result is always in lowercase.

A character string converted to dot-separated format. The result is always in lowercase and
does not start with a dot.
}
\description{
Converts a string from camelCase or underscore_separated format to dot.separated.name format.
Inserts dots before each uppercase letter (except if it's the first character) or replaces underscores with dots,
and then converts the entire string to lowercase.

Converts a string from camelCase or underscore_separated format to dot.separated.name format.
Inserts dots before each uppercase letter (except if it's the first character) or replaces underscores with dots,
and then converts the entire string to lowercase. The output will not start with a dot.
}
\examples{
toDotSeparated("plotMetadataCorHeatMap")
toDotSeparated("plot_Metadata_Cor_HeatMap")
toDotSeparated("plotMetadataCorHeatMap")
toDotSeparated("plot_Metadata_Cor_HeatMap")
}
