% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{toUnderscoreSeparated}
\alias{toUnderscoreSeparated}
\title{Convert a String to underscore_separated Format}
\usage{
toUnderscoreSeparated(input_string, toclipboard = TRUE)
}
\arguments{
\item{input_string}{A character string in camelCase, dot-separated format, or a combination of both.
There is no default value for this parameter; a string must be provided.}

\item{toclipboard}{Copy to clipboard? Default: TRUE}
}
\value{
A character string converted to underscore_separated format.
}
\description{
This function converts a string from camelCase or dot-separated format to an underscore-separated format.
It can handle strings that are a combination of camelCase and dot-separated formats. The function replaces
dots with underscores and inserts an underscore before any uppercase letter that follows a lowercase letter.
It then converts all characters to lowercase.
}
\examples{
toUnderscoreSeparated("plot.Metadata.cor.heatMap")
toUnderscoreSeparated("plotMetadataCorHeatMap")
toUnderscoreSeparated("plot.metadataCor.heatMap")

}
