% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stringendo.R
\name{warnifnot}
\alias{warnifnot}
\title{Issue warnings if conditions are not TRUE}
\usage{
warnifnot(...)
}
\arguments{
\item{...}{Logical conditions to be checked. Each condition must evaluate to a logical vector.
Default: none.}
}
\value{
The function returns \code{invisible()}, meaning it does not produce any visible output. It
issues warnings for each condition that evaluates to \code{FALSE} or contains \code{NA}.
}
\description{
The \code{warningifnot()} function checks whether each condition passed to it is \code{TRUE}. If any
condition is not met, a warning is issued but the execution continues. This is similar to
\code{stopifnot()}, which throws an error and halts execution, but \code{warningifnot()} only issues a
warning, allowing the program to proceed.
}
\examples{
x <- 5
y <- 10

# Example: This will issue a warning because x > y is FALSE
warnifnot(x > y, y > 5)

}
